+++ examples/custom_style.rs
@@ -42,7 +42,7 @@ fn main() -> core::fmt::Result {
             "y",
         );
 
-        let x = (0..50).map(|x| (x as f32 / 50.0) * 10.0);
+        let x = (0..50).map(|x| (x as f64 / 50.0) * 10.0);
 
         s.line("cos", x.clone().map(|x| [x, x.cos()]));
         s.histogram("sin-10", x.clone().step_by(3).map(|x| [x, x.sin() - 10.]));
